solution to the allow the puppy agent to build
